requirements.txt
python-igraph==0.9.10
pycairo==1.21.0
